pyproject.toml
[tool.poetry]
name = "fake-analytics"
version = "0.1.0"
description = "A tool to fake website analytics"
authors = ["Jules"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
playwright = "^1.56.0"
python-dotenv = "^1.2.1"
faker = "^22.0.0"
click = "^8.3.1"
rich = "^13.7.0"
pytz = "^2024.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
ruff = "^0.6.0"
pre-commit = "^3.7.0"
sphinx = "^7.0.0"
sphinx-rtd-theme = "^2.0.0"
sphinx-autodoc-typehints = "^1.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fake-analytics = "fake_analytics.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-n",
    "auto",
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src/fake_analytics",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "-s"
]
markers = [
    "unit: Unit tests for individual modules",
    "integration: Integration tests for module interactions",
    "e2e: End-to-end functional tests",
    "slow: Tests that take a long time to run",
    "network: Tests that require network connectivity",
]
asyncio_mode = "auto"
# Log configuration
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
# Optional: log to file (uncomment to enable)
# log_file = "pytest.log"
# log_file_level = "DEBUG"
# log_file_format = "%(asctime)s [%(levelname)8s] %(filename)s:%(lineno)d: %(message)s"
# log_file_date_format = "%Y-%m-%d %H:%M:%S"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
# Set the maximum line length
line-length = 100

# Exclude directories
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
    ".pytest_cache",
    ".coverage",
    "htmlcov",
    "chrome_profile",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "DTZ", "T10", "ERA", "PD", "PGH", "NPY", "RUF"]
ignore = [
    "E501",  # Line too long (handled by formatter)
    "B008",  # Do not perform function calls in argument defaults
    "PLR0913",  # Too many arguments
    "PLR2004",  # Magic value used in comparison
    "PLE1205",  # Too many arguments for logging format (custom logger, not stdlib logging)
]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Use 4 spaces for indentation
skip-magic-trailing-comma = false
# Line ending style
line-ending = "auto"

[tool.ruff.lint.isort]
# Use a known import section ordering
known-first-party = ["fake_analytics"]
